// projects/gym-library/src/lib/models/environment.model.ts
export interface FirebaseConfig {
  apiKey: string;
  authDomain: string;
  projectId: string;
  storageBucket: string;
  messagingSenderId: string;
  appId: string;
  measurementId?: string;
}

export interface Environment {
  production: boolean;
  useEmulator?: boolean;
  firebase: FirebaseConfig;
}